!parsing struct is ok.
